+++ quiz_quest/lib/app/features/quiz_pages/music_quiz_pages/medium_music_quiz_page/medium_question_quiz_page.dart
@@ -26,7 +26,9 @@ import 'package:quiz_quest/app/features/quiz_pages/history_quiz_pages/easy_histo
 import 'package:quiz_quest/app/features/quiz_pages/music_quiz_pages/cubit/music_cubit.dart';
 import 'package:quiz_quest/app/features/quiz_pages/music_quiz_pages/easy_music_quiz_page/easy_music_lost_life_page.dart';
 import 'package:quiz_quest/app/features/quiz_pages/music_quiz_pages/easy_music_quiz_page/resume_easy_music_question_quiz_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/music_quiz_pages/medium_music_quiz_page/medium_music_answer_button.dart';
 import 'package:quiz_quest/app/features/quiz_pages/music_quiz_pages/medium_music_quiz_page/medium_music_lost_life_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/music_quiz_pages/medium_music_quiz_page/medium_music_question_widget.dart';
 import 'package:quiz_quest/app/features/quiz_pages/music_quiz_pages/medium_music_quiz_page/resume_medium_music_question_quiz_page.dart';
 import 'package:quiz_quest/app/features/quiz_pages/quiz_countdown_timer/quiz_countdown_timer.dart';
 import 'package:quiz_quest/app/injection_container.dart';
@@ -56,7 +58,8 @@ String threeLives = '❤️❤️❤️';
 String twoLives = ' ❤️❤️';
 String oneLive = ' ❤️';
 
-class MediumQuestionMusicQuizPageState extends State<MediumQuestionMusicQuizPage> {
+class MediumQuestionMusicQuizPageState
+    extends State<MediumQuestionMusicQuizPage> {
   int currentIndex = 0;
   late List currentAnswers;
 
@@ -127,8 +130,7 @@ class MediumQuestionMusicQuizPageState extends State<MediumQuestionMusicQuizPage
 
     return Scaffold(
       body: BlocProvider(
-        create: (context) => getIt<MusicCubit>()
-          ..getMediumMusicCategory(),
+        create: (context) => getIt<MusicCubit>()..getMediumMusicCategory(),
         child: BlocListener<MusicCubit, MusicState>(
           listener: (context, state) async {
             if (state.status == Status.error) {
@@ -246,9 +248,8 @@ class MediumQuestionMusicQuizPageState extends State<MediumQuestionMusicQuizPage
                           isButtonDisabled = true;
                           isTimeUp = true;
                           if (mediumMusicBadAnswers == 3) {
-                            context
-                                .read<MusicCubit>()
-                                .updateMediumMusicPoints(mediumMusicGoodAnswers);
+                            context.read<MusicCubit>().updateMediumMusicPoints(
+                                mediumMusicGoodAnswers);
                             Navigator.of(context).push(
                               MaterialPageRoute(
                                 builder: (context) => MediumMusicLostLifePage(
@@ -281,7 +282,7 @@ class MediumQuestionMusicQuizPageState extends State<MediumQuestionMusicQuizPage
                             return Column(
                               mainAxisAlignment: MainAxisAlignment.spaceAround,
                               children: [
-                                QuestionWidget(
+                                MediumMusicQuestionWidget(
                                   question: musicQuizModel
                                       .results[currentIndex].question,
                                 ),
@@ -327,7 +328,7 @@ class MediumQuestionMusicQuizPageState extends State<MediumQuestionMusicQuizPage
                                 for (int index = 0;
                                     index < currentAnswers.length;
                                     index++) ...[
-                                  AnswerButton(
+                                  MediumMusicAnswerButton(
                                     isTimeUp: isTimeUp,
                                     duration: duration,
                                     isButtonDisabled: (value) {
@@ -445,151 +446,6 @@ class MediumQuestionMusicQuizPageState extends State<MediumQuestionMusicQuizPage
                 ),
               );
             },
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class QuestionWidget extends StatelessWidget {
-  const QuestionWidget({
-    required this.question,
-    super.key,
-  });
-
-  final String question;
-
-  @override
-  Widget build(BuildContext context) {
-    final modifiedQuestion = question
-        .replaceAll('&quot;', '')
-        .replaceAll('&#039;', '')
-        .replaceAll('&aacute;', '')
-        .replaceAll('&ntilde;', '')
-        .replaceAll('&amp;', '')
-        .replaceAll('&rsquo;', '');
-
-    return Center(
-      child: Text(
-        modifiedQuestion,
-        style: GoogleFonts.aBeeZee(
-            fontSize: 24, color: Colors.white, fontWeight: FontWeight.bold),
-        textAlign: TextAlign.center,
-      ),
-    );
-  }
-}
-
-class AnswerButton extends StatefulWidget {
-  AnswerButton({
-    required this.answer,
-    required this.controller,
-    required this.isCorrectAnswer,
-    required this.colorFunction,
-    required this.isButtonClicked,
-    required this.isButtonDisabled,
-    required this.textcolor,
-    required this.index,
-    required this.duration,
-    required this.isTimeUp,
-  });
-
-  final String answer;
-  final CountDownController controller;
-  final bool isCorrectAnswer;
-
-  int duration;
-  bool isTimeUp;
-  Function(Color, int) colorFunction;
-  final Function(bool) isButtonClicked;
-  final Function(bool) isButtonDisabled;
-  Color textcolor;
-  final int index;
-
-  @override
-  State<AnswerButton> createState() => _AnswerButtonState();
-}
-
-class _AnswerButtonState extends State<AnswerButton> {
-  void onPressed() {
-    if (isButtonDisabled) {
-      return;
-    }
-
-    widget.controller.pause();
-
-    if (widget.isCorrectAnswer) {
-      widget.colorFunction(Colors.green, widget.index);
-      mediumMusicGoodAnswers += 1;
-    } else {
-      widget.colorFunction(Colors.red, widget.index);
-      mediumMusicBadAnswers += 1;
-      if (mediumMusicBadAnswers == 3) {
-        context.read<MusicCubit>().updateMediumMusicPoints(mediumMusicGoodAnswers);
-        Navigator.of(context).push(
-          MaterialPageRoute(
-            builder: (context) =>
-                MediumMusicLostLifePage(goodAnswers: mediumMusicGoodAnswers),
-          ),
-        );
-      }
-    }
-    setState(() {
-      widget.textcolor = widget.isCorrectAnswer ? Colors.green : Colors.red;
-    });
-    widget.isButtonClicked(true);
-    widget.isButtonDisabled(true);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final modifiedAnswer = widget.answer
-        .replaceAll('&quot;', '')
-        .replaceAll('&#039;', '')
-        .replaceAll('&aacute;', '')
-        .replaceAll('&ntilde;', '')
-        .replaceAll('&amp;', '')
-        .replaceAll('&rsquo;', '')
-        .replaceAll('&ocirc;', '');
-
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: 20),
-      decoration: BoxDecoration(
-        gradient: const LinearGradient(
-          colors: [
-            Color.fromRGBO(11, 22, 65, 1),
-            Color.fromRGBO(9, 77, 203, 1),
-          ],
-          begin: Alignment.centerLeft,
-          end: Alignment.centerRight,
-        ),
-        borderRadius: const BorderRadius.all(
-          Radius.circular(6.0),
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            spreadRadius: 4,
-            blurRadius: 10,
-            offset: const Offset(0, 3),
-          )
-        ],
-      ),
-      child: ElevatedButton(
-        onPressed: onPressed,
-        style: ElevatedButton.styleFrom(
-          minimumSize: const Size.fromHeight(50),
-          backgroundColor: Colors.transparent,
-          shadowColor: Colors.transparent,
-        ),
-        child: Text(
-          modifiedAnswer,
-          style: GoogleFonts.aBeeZee(
-            fontSize: 24,
-            color: widget.isTimeUp && widget.isCorrectAnswer
-                ? Colors.green
-                : widget.textcolor,
           ),
         ),
       ),
